@@ -4,34 +4,33 @@ using Grpc.Core;
 using Grpc.Net.Client;
 using System;
 using System.Threading.Tasks;
-namespace CasCap
+namespace CasCap;
+
+class Program : ProgramBase
 {
-    class Program : ProgramBase
+    static async Task Main()
     {
-        static async Task Main()
+        await Run();
+        Console.WriteLine("Press any key to exit...");
+        Console.ReadKey();
+    }
+
+    async static Task Run()
+    {
+        using var channel = GrpcChannel.ForAddress("https://localhost:5001");
+        var client = new Markets.MarketsClient(channel);
+
+        //var cts = new CancellationTokenSource(TimeSpan.FromSeconds(20));
+        using var streamingCall = client.GetTickStream(new Empty()/*, cancellationToken: cts.Token*/);
+
+        try
         {
-            await Run();
-            Console.WriteLine("Press any key to exit...");
-            Console.ReadKey();
+            await foreach (var tick in streamingCall.ResponseStream.ReadAllAsync(/*cancellationToken: cts.Token*/))
+                Display(tick);
         }
-
-        async static Task Run()
+        catch (RpcException ex) when (ex.StatusCode == StatusCode.Cancelled)
         {
-            using var channel = GrpcChannel.ForAddress("https://localhost:5001");
-            var client = new Markets.MarketsClient(channel);
-
-            //var cts = new CancellationTokenSource(TimeSpan.FromSeconds(20));
-            using var streamingCall = client.GetTickStream(new Empty()/*, cancellationToken: cts.Token*/);
-
-            try
-            {
-                await foreach (var tick in streamingCall.ResponseStream.ReadAllAsync(/*cancellationToken: cts.Token*/))
-                    Display(tick);
-            }
-            catch (RpcException ex) when (ex.StatusCode == StatusCode.Cancelled)
-            {
-                Console.WriteLine("Stream cancelled.");
-            }
+            Console.WriteLine("Stream cancelled.");
         }
     }
 }
