@@ -3,62 +3,61 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-namespace CasCap
+namespace CasCap;
+
+//https://stackoverflow.com/questions/5852863/fixed-size-queue-which-automatically-dequeues-old-values-upon-new-enques
+[Serializable]
+[DebuggerDisplay("Count = {" + nameof(Count) + "}, Limit = {" + nameof(Limit) + "}")]
+public class FixedSizedQueue<T> : IReadOnlyCollection<T>
 {
-    //https://stackoverflow.com/questions/5852863/fixed-size-queue-which-automatically-dequeues-old-values-upon-new-enques
-    [Serializable]
-    [DebuggerDisplay("Count = {" + nameof(Count) + "}, Limit = {" + nameof(Limit) + "}")]
-    public class FixedSizedQueue<T> : IReadOnlyCollection<T>
+    private readonly Queue<T> _queue = new();
+    private readonly object _lock = new();
+
+    public int Count { get { lock (_lock) { return _queue.Count; } } }
+    public int Limit { get; }
+
+    public FixedSizedQueue(int limit)
     {
-        private readonly Queue<T> _queue = new();
-        private readonly object _lock = new();
+        if (limit < 1)
+            throw new ArgumentOutOfRangeException(nameof(limit));
 
-        public int Count { get { lock (_lock) { return _queue.Count; } } }
-        public int Limit { get; }
+        Limit = limit;
+    }
 
-        public FixedSizedQueue(int limit)
+    public FixedSizedQueue(IEnumerable<T> collection)
+    {
+        if (collection is null || !collection.Any())
+            throw new ArgumentException("Can not initialize the Queue with a null or empty collection", nameof(collection));
+
+        _queue = new Queue<T>(collection);
+        Limit = _queue.Count;
+    }
+
+    public void Enqueue(T obj)
+    {
+        lock (_lock)
         {
-            if (limit < 1)
-                throw new ArgumentOutOfRangeException(nameof(limit));
+            _queue.Enqueue(obj);
 
-            Limit = limit;
+            while (_queue.Count > Limit)
+                _queue.Dequeue();
         }
+    }
 
-        public FixedSizedQueue(IEnumerable<T> collection)
-        {
-            if (collection is null || !collection.Any())
-                throw new ArgumentException("Can not initialize the Queue with a null or empty collection", nameof(collection));
+    public void Clear()
+    {
+        lock (_lock)
+            _queue.Clear();
+    }
 
-            _queue = new Queue<T>(collection);
-            Limit = _queue.Count;
-        }
+    public IEnumerator<T> GetEnumerator()
+    {
+        lock (_lock)
+            return new List<T>(_queue).GetEnumerator();
+    }
 
-        public void Enqueue(T obj)
-        {
-            lock (_lock)
-            {
-                _queue.Enqueue(obj);
-
-                while (_queue.Count > Limit)
-                    _queue.Dequeue();
-            }
-        }
-
-        public void Clear()
-        {
-            lock (_lock)
-                _queue.Clear();
-        }
-
-        public IEnumerator<T> GetEnumerator()
-        {
-            lock (_lock)
-                return new List<T>(_queue).GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
+    IEnumerator IEnumerable.GetEnumerator()
+    {
+        return GetEnumerator();
     }
 }
